Ho roughest draft (#134)
* adding first brain dump for intro/motivation

* into -- Friday 8/7 edits

* Aug 13 edits: intro and starting graph overview

* more on latent confounding

* latent confounding format

* adding break lines

* more intro text and fixes

* more overview of causal graphs, and more on latent confounding

* grabbing updated compilation script

* adding citations, and figure references

* more on causal graph overview with equations

* adding figures folders and files, and fixing figure references

* adding citation for pearl 2000, and fixing figure captions

* text edits, adding breaks after lines, and figure adjustments
