+++ article/images/simple-graph
@@ -1,0 +1,8 @@
+digraph {
+	Z [label=Z shape=ellipse]
+	Y [label=Y shape=ellipse]
+	X [label=X shape=ellipse]
+	X -> Z
+	X -> Y
+	Z -> Y
+}
