+++ article/images/simple-graph-confounded
@@ -1,0 +1,11 @@
+digraph {
+	Z [label=Z shape=ellipse]
+	Y [label=Y shape=ellipse]
+	U [label=U shape=ellipse]
+	X [label=X shape=ellipse]
+	X -> Z
+	X -> Y
+	Z -> Y
+	U -> Z
+	U -> Y
+}
