+++ article/main.bib
@@ -225,6 +225,13 @@
   publisher={Oxford University Press}
 }
 
+@book{pearl_causality_2000,
+  title={Causality: Models, Reasoning, and Inference},
+  author={Pearl, Judea},
+  year={2000},
+  publisher={Cambridge University Press}
+}
+
 @article{pitchforth_2013_proposed,
   title={A proposed validation framework for expert elicited Bayesian Networks},
   author={Pitchforth, Jegar and Mengersen, Kerrie},
@@ -383,4 +390,3 @@
   pages={3414--3425},
   year={2020},
   organization={PMLR}
-}
