+++ article/sections/_1_intro_generic.tex
@@ -1,2 +1,54 @@
 \section{Introduction}
 \label{sec:intro}
+
+The following chapter is concerned with the existing disconnect between the 
+fields of travel demand modeling and causal inference. 
+More specifically, the chapter is motivated by the current lack of use of methods and findings from 
+the causal inference literature in travel demand modeling. 
+
+More often than not, the development of behavioral models in transportation, 
+specifically transportation demand models, is driven by the need to evaluate 
+the impact external interventions, in the form of alternative policies, on a 
+certain outcome of interest. 
+These questions that transportation demand models 
+are built to answer are causal by definition: we are interested in how the 
+system reacts to \textit{external} interventions. 
+Yet, when those models are 
+developed, there is very little consideration given to causality, and when 
+causal concepts are accounted for, the process is done implicitly without a 
+formal framework. 
+
+While the field of transportation demand modeling could benefit greatly from 
+incorporating causal inference techniques, there are barriers that have made 
+this integration slower than what one would hope. 
+These barriers stem from the difference between the types of problems transportation demand modelers deal with and those that are typically studied in the causal inference literature. 
+Perhaps the main fundemental difference is that demand modelers 
+are typically trying to forecast the impacts of policies that haven't been 
+implemented or seen before, which requires additional work and a change to the
+typical causal modeling workflow in order to translate a given policy 
+(treatment) into a set of characteristics and variables that exist in the
+data and system at hand (please refer to \citet{brathwaite_2018_causal} for a 
+more thorough discussion of those barriers). 
+While those barriers make the problem of demand modelers harder, there is still a lot to gain from 
+incorporating causal inference techniques where appropriate, and to contribute
+in turn where the literature lacks.  
+
+The relevance of this topic now is motivated by the significant boost in the causal inference literature recently, both in the potential outcomes framework and the causal graphical approach.
+The goal of this chapter is to formalize a framework for approaching 
+transportation demand modeling problems from a causal perspective. 
+We will draw heavily on the use of directed acyclic graphs (DAGs) formalized by \citet{pearl_causality_2000} as a means of representing the modeler's knowledge and assumptions about a given problem. 
+The chapter will provide an overview of DAGs, the 
+testable implications that come with one's causal representation, the main 
+tests that one could do to falsify/justify a given causal graph, and then how 
+to use a causal graph to estimate the causal relationships of interest. 
+We will demonstrate the use of this framework through simulations, where we 
+clearly show the benefits and implications of this approach as opposed to 
+traditional approaches. 
+The last part of this chapter deals with the more complicated issue of latent 
+confounding, where one variable confounds two or more variables in the causal 
+graph. This type of confounding creates variations in the outcome variable that are not caused by 
+the confounded variables but are correlated by it, which biases the estimated 
+causal effects of those variables if nothing is done to account for the 
+confounding. 
+We focus on a recent technique to deal with latent confounding suggested by \citet{wang_2019_blessings} to address unobsrved confounding in tranportation applications. 
+We describe this problem in more details in section 7.
