+++ notebooks/working/factor_models.py
@@ -1,0 +1,238 @@
+# +
+import pandas as pd
+import numpy as np
+import matplotlib.pyplot as plt
+from causalgraphicalmodels import CausalGraphicalModel, StructuralCausalModel
+import pylogit
+from collections import OrderedDict
+import pylogit as cm
+from functools import reduce
+import statsmodels.api as sm
+import statsmodels.formula.api as smf
+from math import ceil
+
+from IPython import display
+import seaborn as sns
+
+import numpy as np
+import numpy.random as npr
+import pandas as pd
+import tensorflow as tf
+import tensorflow_probability as tfp
+import statsmodels.api as sm
+
+from tensorflow_probability import edward2 as ed
+from sklearn.datasets import load_breast_cancer
+from pandas.plotting import scatter_matrix
+from scipy import sparse, stats
+from sklearn.model_selection import train_test_split
+from sklearn.linear_model import LogisticRegression 
+from sklearn.metrics import confusion_matrix, classification_report, roc_auc_score, roc_curve
+
+import matplotlib
+matplotlib.rcParams.update({'font.sans-serif' : 'Helvetica',
+                            'axes.labelsize': 10,
+                            'xtick.labelsize' : 6,
+                            'ytick.labelsize' : 6,
+                            'axes.titlesize' : 10})
+import matplotlib.pyplot as plt
+
+import seaborn as sns
+color_names = ["windows blue",
+               "amber",
+               "crimson",
+               "faded green",
+               "dusty purple",
+               "greyish"]
+colors = sns.xkcd_palette(color_names)
+sns.set(style="white", palette=sns.xkcd_palette(color_names), color_codes = False)
+
+
+from factor_models import *
+
+
+
+
+"""
+References: Blei et al 2018 (https://github.com/blei-lab/deconfounder_tutorial) 
+"""
+
+
+
+def confounder_ppca(X, latent_dim, holdout_portion):
+    """
+    Function to estimate a substitute confounder using PPCA.
+    Adopted from the deconfounder_tutorial.ipynb
+    https://github.com/blei-lab/deconfounder_tutorial
+    
+    Args:
+        X: A numpy array or pandas dataframe of the original covariates
+           dimension: (n x m)
+        latent_dim: The number of latend factors to be estimated
+        holdout_portion: Fraction of the data to be used as holdout
+        
+    Returns:
+        w_mean_inferred: (latent_dim x n) matrix
+        w_std_inferred: (latent_dim x n) matrix
+        z_mean_inferred: mean of substitute confounder
+                         dimension (n x latend_dim)
+        z_std_inferred: std of substitute confounder
+                         dimension (n x latend_dim)
+                         
+        x_vad: (nxm) matrix with the heldout entries only and 0 elsewhere
+        holdout_mask: sparse (nxm) matrix with 1 on the heldout entries and
+                      0 elsewhere, s.t x_vad = X*holdout_mask
+        holdout_rows: row indeces of the heldout entries
+    
+    
+    """
+    
+    
+    num_datapoints, data_dim = X.shape
+
+    holdout_portion = holdout_portion
+    n_holdout = int(holdout_portion * num_datapoints * data_dim)
+
+    holdout_row = np.random.randint(num_datapoints, size=n_holdout)
+    holdout_col = np.random.randint(data_dim, size=n_holdout)
+    holdout_mask = (sparse.coo_matrix((np.ones(n_holdout), \
+                                (holdout_row, holdout_col)), \
+                                shape = X.shape)).toarray()
+
+    holdout_subjects = np.unique(holdout_row)
+
+    x_train = np.multiply(1-holdout_mask, X)
+    x_vad = np.multiply(holdout_mask, X)
+
+    def ppca_model(data_dim, latent_dim, num_datapoints, stddv_datapoints):
+        w = ed.Normal(loc=tf.zeros([latent_dim, data_dim]),
+                    scale=tf.ones([latent_dim, data_dim]),
+                    name="w")  # parameter
+        z = ed.Normal(loc=tf.zeros([num_datapoints, latent_dim]),
+                    scale=tf.ones([num_datapoints, latent_dim]), 
+                    name="z")  # local latent variable / substitute confounder
+        x = ed.Normal(loc=tf.multiply(tf.matmul(z, w), 1-holdout_mask),
+                    scale=stddv_datapoints * tf.ones([num_datapoints, data_dim]),
+                    name="x")  # (modeled) data
+        return x, (w, z)
+
+    log_joint = ed.make_log_joint_fn(ppca_model)
+
+    latent_dim = latent_dim
+    stddv_datapoints = 0.1
+
+    model = ppca_model(data_dim=data_dim,
+                       latent_dim=latent_dim,
+                       num_datapoints=num_datapoints,
+                       stddv_datapoints=stddv_datapoints)
+
+    def variational_model(qw_mean, qw_stddv, qz_mean, qz_stddv):
+        qw = ed.Normal(loc=qw_mean, scale=qw_stddv, name="qw")
+        qz = ed.Normal(loc=qz_mean, scale=qz_stddv, name="qz")
+        return qw, qz
+
+
+    log_q = ed.make_log_joint_fn(variational_model)
+
+    def target(w, z):
+        """Unnormalized target density as a function of the parameters."""
+        return log_joint(data_dim=data_dim,
+                       latent_dim=latent_dim,
+                       num_datapoints=num_datapoints,
+                       stddv_datapoints=stddv_datapoints,
+                       w=w, z=z, x=x_train)
+
+    def target_q(qw, qz):
+        return log_q(qw_mean=qw_mean, qw_stddv=qw_stddv,
+                   qz_mean=qz_mean, qz_stddv=qz_stddv,
+                   qw=qw, qz=qz)
+
+
+    qw_mean = tf.Variable(np.ones([latent_dim, data_dim]), dtype=tf.float32)
+    qz_mean = tf.Variable(np.ones([num_datapoints, latent_dim]), dtype=tf.float32)
+    qw_stddv = tf.nn.softplus(tf.Variable(-4 * np.ones([latent_dim, data_dim]), dtype=tf.float32))
+    qz_stddv = tf.nn.softplus(tf.Variable(-4 * np.ones([num_datapoints, latent_dim]), dtype=tf.float32))
+
+    qw, qz = variational_model(qw_mean=qw_mean, qw_stddv=qw_stddv,
+                               qz_mean=qz_mean, qz_stddv=qz_stddv)
+
+
+    energy = target(qw, qz)
+    entropy = -target_q(qw, qz)
+
+    elbo = energy + entropy
+
+
+    optimizer = tf.train.AdamOptimizer(learning_rate = 0.05)
+    train = optimizer.minimize(-elbo)
+
+    init = tf.global_variables_initializer()
+
+    t = []
+
+    num_epochs = 500
+
+    with tf.Session() as sess:
+        sess.run(init)
+
+        for i in range(num_epochs):
+            sess.run(train)
+            if i % 5 == 0:
+                t.append(sess.run([elbo]))
+
+            w_mean_inferred = sess.run(qw_mean)
+            w_stddv_inferred = sess.run(qw_stddv)
+            z_mean_inferred = sess.run(qz_mean)
+            z_stddv_inferred = sess.run(qz_stddv)
+
+    print("Inferred axes:")
+    print(w_mean_inferred)
+    print("Standard Deviation:")
+    print(w_stddv_inferred)
+
+    plt.plot(range(1, num_epochs, 5), t)
+    plt.show()
+
+    def replace_latents(w, z):
+
+        def interceptor(rv_constructor, *rv_args, **rv_kwargs):
+            """Replaces the priors with actual values to generate samples from."""
+            name = rv_kwargs.pop("name")
+            if name == "w":
+                rv_kwargs["value"] = w
+            elif name == "z":
+                rv_kwargs["value"] = z
+            return rv_constructor(*rv_args, **rv_kwargs)
+
+        return interceptor
+    
+    return [w_mean_inferred, w_stddv_inferred, z_mean_inferred, z_stddv_inferred], x_vad, holdout_mask, holdout_row
+
+
+def ppca_model(data_dim, latent_dim, num_datapoints, stddv_datapoints, holdout_mask):
+    w = ed.Normal(loc=tf.zeros([latent_dim, data_dim]),
+                scale=tf.ones([latent_dim, data_dim]),
+                name="w")  # parameter
+    z = ed.Normal(loc=tf.zeros([num_datapoints, latent_dim]),
+                scale=tf.ones([num_datapoints, latent_dim]), 
+                name="z")  # local latent variable / substitute confounder
+    x = ed.Normal(loc=tf.multiply(tf.matmul(z, w), 1-holdout_mask),
+                scale=stddv_datapoints * tf.ones([num_datapoints, data_dim]),
+                name="x")  # (modeled) data
+    return x, (w, z)
+
+
+
+
+def replace_latents(w, z):
+
+    def interceptor(rv_constructor, *rv_args, **rv_kwargs):
+        """Replaces the priors with actual values to generate samples from."""
+        name = rv_kwargs.pop("name")
+        if name == "w":
+            rv_kwargs["value"] = w
+        elif name == "z":
+            rv_kwargs["value"] = z
+        return rv_constructor(*rv_args, **rv_kwargs)
+
+    return interceptor
